+++ edx_ace/recipient.py
@@ -13,8 +13,8 @@ class Recipient(MessageAttributeSerializationMixin):
     The target for a message.
 
     Arguments:
-        username (str): The username of the intended recipient.
+        lms_user_id (int): The LMS user ID of the intended recipient.
         email_address (str): The email address of the intended recipient. Optional.
     """
-    username = attr.ib()
+    lms_user_id = attr.ib()
     email_address = attr.ib(default=None)