+++ edx_ace/tests/channel/test_file_email.py
@@ -40,7 +40,7 @@ class TestFilesEmailChannel(TestCase):
                     'from_address': 'bulk@example.com',
                     PATH_OVERRIDE_KEY: f.name,
                 },
-                recipient=Recipient(username='Robot', email_address='mr@robot.io'),
+                recipient=Recipient(lms_user_id=123, email_address='mr@robot.io'),
             )
 
             channel.deliver(message, rendered_message)